chore(refactor): use all utils from the root file in generators
use all utils from root file in generators
