+++ packages/generators/init-generator.ts
@@ -6,13 +6,18 @@ import * as path from "path";
 import { getPackageManager } from "@webpack-cli/utils/package-manager";
 import { Confirm, Input, List } from "@webpack-cli/webpack-scaffold";
 
-import { getDefaultOptimization } from "./utils/webpackConfig";
+import {
+	getDefaultOptimization,
+	LangType,
+	langQuestionHandler,
+	tooltip,
+	generatePluginName,
+	Loader,
+	StylingType,
+	styleQuestionHandler,
+	entryQuestions
+} from "./utils";
 import { WebpackOptions } from "./types";
-import entryQuestions from "./utils/entry";
-import langQuestionHandler, { LangType } from "./utils/languageSupport";
-import styleQuestionHandler, { Loader, StylingType } from "./utils/styleSupport";
-import tooltip from "./utils/tooltip";
-import { generatePluginName } from "./utils/plugins";
 
 /**
  *
